Generate the save path for the current series
Implement BaseSeriesDirectory to store the base portion of the generated
directory. Using the series' name, append it to the base portion. Set
the lbSeriesDestination's Text property to the created path. If the
path already exists, set the checked value of the corresponding RadioButton
to True.

Refactor code in FormMain.cs