Add series destination checkbox
Remove the Label control displaying the destination and reintroduce
the TextBox. In addition, allow the user to change the save destination
by typing into the text control. Perform validation on the input text to
make sure the path is valid.

In place of the RadioButton controls, use a CheckBox to specify whether
the chapters are to be downloaded to the series' directory.

Give the ToolTip a more descriptive name.