Use the 'txtSaveTo' text for the series base
Use the current text in the 'txtSaveTo' TextBox as the base directory
for the series folder instead of 'My Documents'.

When choosing to save the chapters in the series folder, the base
directory defaults to 'My Documents' if 'BaseSeriesDestination' hasn't
been set. Users may expect it to be saved in a sub-directory of
the current save location and not My Documents.