Fix KissManga URI decryption error
Attempt a fix by removing the code for the "CryptoJS" search which
prevents the decryption function from being called.

The KissManga website appears to have been updated, removing previous
values which were required by the plugin for decryption. If the plugin
couldn't find the value, it would throw an error.