Check for and create the CBZ destination directory
An error occurs if the user selects CBZ as the output and the directory
specified in 'task.SaveToFolder' doesn't exist. The bug was most-likely
introduced with the 'Save in Series Folder' feature. Windows
automatically creates all missing directories in the path when creating
a new directory.

Resolve the bug by checking for the existence of the directory and
manually creating it before attempting to create the CBZ file.

Referencing issue #94